requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.1
